testing adjustment

+++ debiasm/torch_functions.py
@@ -227,7 +227,7 @@ def DEBIASM_train_and_pred(X_train,
     baseline_mod.fit(rescale( X_train[:, 1:]), y_train)
     
         
-    model = PL_DEBIASM(X = torch.tensor( np.vstack((X_train, X_val)) ),
+    model = PL_DEBIASM(X = torch.tensor( np.vstack((X_train, X_val)).astype(float) ),
                        batch_sim_strength = batch_sim_strength,
                        input_dim = X_train.shape[1]-1, 
                        num_classes = 2, 
@@ -473,7 +473,7 @@ def DEBIASM_train_and_pred_log_additive(X_train,
     baseline_mod.fit( X_train[:, 1:], y_train)
 
     
-    model = PL_DEBIASM_log_additive(X = torch.tensor( np.vstack((X_train, X_val)) ),
+    model = PL_DEBIASM_log_additive(X = torch.tensor( np.vstack((X_train, X_val)).astype(float) ),
                                     batch_sim_strength = batch_sim_strength,
                                     input_dim = X_train.shape[1]-1, 
                                     num_classes = 2,
